@@ -5,18 +5,21 @@
   "proxy": "http://localhost:3000",
   "devDependencies": {},
   "dependencies": {
-    "antd": "^3.2.1",
+    "antd": "^4.22.8",
     "forever": "latest",
-    "cronstrue": "^1.1.0",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "react-router-dom": "^4.1.1",
-    "redux": "^3.6.0",
-    "react-redux": "^5.0.2",
-    "express": "^4.15.2",
-    "history": "^4.6.3",
-    "react-fa": "^4.2.0",
-    "react-scripts": "1.0.11"
+    "cronstrue": "^2.11.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "redux": "^4.2.0",
+    "react-redux": "^8.0.2",
+    "express": "^4.18.1",
+    "history": "^5.3.0",
+    "react-fontawesome": "^1.7.1",
+    "react-scripts": "5.0.1",
+    "domhandler": "^5.0.3",
+    "domelementtype": "2.3.0",
+    "@ant-design/icons": "4.7.0"
   },
   "scripts": {
     "start": "react-scripts start",
